requirements.txt
tensorflow
Flask==3.0.2
Werkzeug==3.0.1
Pillow==10.2.0
numpy==1.26.4
requests==2.31.0
matplotlib==3.8.3
